Moves almost all environment variables to a .env file
This removes almost all of the build environment variables from the
Dockerfile and puts them into a separate file that is attached to the
build container when the container is started (docker run command
--env-file option). That option is not REQUIRED on when the build
container is required since that is the only source for most of the
configuration parameters for OpenPegasus.

The only environment variables left in the Dockerfile are the default
values for the directories and path in the build container which should
not change, the OpenPegasus platform which must be 64 bit linux for
ubuntu, and the git configuration which may be overridden by the .env
file.