Adds variables to Makefile to allow autorun for the the run-build-image and run-server-image targets
Provides new variables BLDDEVMODE and RUNDEVMODE.  Each of these
variables has the options true, false, or not set.

The true option sets the corresponding container to start with bash. The
false option always starts the container with it default operation.

For the build image, the default is to start with bash and the user does
the build manually with the make targets.

For the run container, the default is to start the container with the
OpenPegasus server running.

These options are documented in the Makefile