Docker file changes
Add a number of network tools to the build image to be able to run
ipconfig, etc. from the build container.

Add COPY statement to copy the .supplementary_run_files to the build
image

Change the statement that gets the image tag to use the text from the
version.txt file